port to new branch
I couldn't get things working in the home-page branch. So I started a new branch to be a better clean slate. All of my work is in there and should behave as it did in the other branch, but as of this commit, I intentionally didn't add the functionality to the app yet. The files are there with no way to navigate to them. Adding navigation should come in future commits.